homework04/requirements.txt
requests
python-dateutil
xmltodict==0.12.0

